Run the chorus and delay in double too
Same divergence as the reverb tank, in the other two send effects.  Upstream
holds its chorus and delay state in JavaScript numbers and only narrows when a
sample lands in a Float32Array delay buffer; the port held the state itself in
floats and rounded at every step of a feedback path.

Widen both: the chorus's LFO phase and rate, its pre-lowpass state, depth,
delay and feedback, the read positions and their interpolation, and the
mixdown; and the delay's pre-lowpass, its per-line feedback and gain, and the
sample it circulates.  Buffers stay float32 on both sides and the writes narrow
on store, so the port keeps exactly what upstream keeps.

The chorus axis was 80 dB and now reads 94.0 flat across send level.  On the
corpus, gs_chorus_level goes from 25.35 dB to 60.41 and gs_chorus_rate from
19.21 to 23.50, with nothing else moving.  The delay axis was already at the
level the reverb fix lifted everything to, so its own change is not separately
visible here, but it had the same defect and would have shown on a probe that
drove it harder.

Every grid axis outside the two repeated-note ones now sits at 92 dB or better.
Those two are upstream's voice recycling inheriting state across notes, which
is not ours to fix -- see the patch drafted for upstream.

+++ spessasynth_core/include/spessasynth/synthesizer/dsp/chorus.h
@@ -32,23 +32,23 @@ typedef struct {
 typedef struct {
 	SS_ChorusParams parameters;
 	unsigned int maxBufferSize;
-	float preLPFfc;
-	float preLPFa;
-	float preLPFz;
+	double preLPFfc;
+	double preLPFa;
+	double preLPFz;
 	float *leftDelayBuffer;
 	float *rightDelayBuffer;
 	float sampleRate;
-	float phase;
+	double phase;
 	unsigned int write;
-	float gain;
-	float reverbGain;
-	float delayGain;
+	double gain;
+	double reverbGain;
+	double delayGain;
 	/* Fractional sample counts: the read position is interpolated, so
 	 * rounding these to whole samples quantizes the modulation itself. */
-	float depthSamples;
-	float delaySamples;
-	float rateInc;
-	float feedbackGain;
+	double depthSamples;
+	double delaySamples;
+	double rateInc;
+	double feedbackGain;
 } SS_Chorus;
 
 SS_Chorus SPESSASYNTH_EXPORTS *ss_chorus_create(float sampleRate, int maxBufferSize);
